refac(leaflet): make code easier to maintain
